split presentation dto to requests and responses
