+++ src/Application/Service/CategoryService.php
@@ -2,8 +2,8 @@
 
 namespace App\Application\Service;
 
-use App\Presenstation\Dto\CategoryListItem;
-use App\Presenstation\Dto\CategoryListResponse;
+use App\Presenstation\Response\CategoryListItem;
+use App\Presenstation\Response\CategoryListResponse;
 use App\Infrastructure\Repository\CategoryRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Serializer\SerializerInterface;
