+++ src/Presentation/Docs/CommentControllerInterface.php
@@ -6,9 +6,9 @@ namespace App\Presentation\Docs;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
 use OpenApi\Annotations as OA;
-use App\Presenstation\Dto\CreateCommentRequest;
+use App\Presenstation\Request\CreateCommentRequest;
 use Symfony\Component\HttpKernel\Attribute\MapRequestPayload;
-use App\Presenstation\Dto\UpdateCommentRequest;
+use App\Presenstation\Request\UpdateCommentRequest;
 
 interface CommentControllerInterface {
