+++ src/Presentation/Request/PaymentRequest.php
@@ -1,8 +1,6 @@
-<?php
+<?php declare(strict_types=1);
 
-declare(strict_types=1);
-
-namespace App\Presenstation\Dto;
+namespace App\Presenstation\Request;
 
 final class PaymentRequestDto 
 {
