+++ src/Presentation/Response/CategoryListResponse.php
@@ -1,9 +1,9 @@
-<?php
+<?php declare(strict_types=1);
 
-namespace App\Presenstation\Dto;
+namespace App\Presenstation\Response;
 
 
-class CategoryListResponse
+final class CategoryListResponse
 {
     /**
      * @var CategoryListItem[]
